Remove pip-deps-tree command
