--- artifactory/utils/pip/dependencies/testsdata/requirements.txt
@@ -1,28 +1,0 @@
-pk1._test>=1.0,<=2.0
-git+https://myvcs.com/some_dependency@sometag#egg=pkg2_test
-git://git.example.com/MyProject#egg=pkg3
-hg+http://hg.myproject.org/MyProject#egg=pkg4
-    pkg5[something]
-PkG6<1.3
-pkg7==1.0.1
-pkg8 == 3.0.3
-pk-g9
-git+https://git.repo/some_repo.git#egg=pkg10&subdirectory=subdir_path
--e git+https://git.repo/some_repo.git#egg=pkg11&subdirectory=subdir_path
-/Users/barb/pkg12 # This should output a failure message in log
-# This line shouldn't be parsed
-git+https://myvcs.com/some_dependency@sometag#egg=pkg13&subdirectory \
- --- This line shouldn't be parsed 2
-pkg14 #blabla\\\
-pkg15 \
-pkg16
-pk\
-g17
-pkg18\
-# Not included
-pkg19 #comment\\\
-    # pkg20 \
-pkg21
-pkg22 #comment\\\
-   pkg23    # comment \
-pkg24
