+++ artifactory/utils/pip/setuppy.go
@@ -1,91 +1,20 @@
-package dependencies
+package pip
 
 import (
 	"errors"
 	"fmt"
 	gofrogcmd "github.com/jfrog/gofrog/io"
-	"github.com/jfrog/jfrog-cli-go/artifactory/utils/pip"
-	"github.com/jfrog/jfrog-client-go/artifactory/buildinfo"
 	"github.com/jfrog/jfrog-client-go/utils"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
 	"io/ioutil"
 	"os"
-	"strings"
-	"sync"
 )
 
-// Dependencies extractor for setup.py
-type setupExtractor struct {
-	allDependencies      map[string]*buildinfo.Dependency
-	childrenMap          map[string][]string
-	rootDependencies     []string
-	setuppyFilePath      string
-	pythonExecutablePath string
-	Pkg                  string
-	once                 sync.Once
-}
-
-func NewSetupExtractor(setuppyFilePath, pythonExecutablePath string) Extractor {
-	// Create new extractor.
-	return &setupExtractor{setuppyFilePath: setuppyFilePath, pythonExecutablePath: pythonExecutablePath}
-}
-
-func (extractor *setupExtractor) Extract() error {
-	// Get installed packages tree.
-	environmentPackages, err := BuildPipDependencyMap(extractor.pythonExecutablePath)
-	if err != nil {
-		return err
-	}
-
-	pkgName, err := extractor.PackageName()
-	if err != nil {
-		return err
-	}
-
-	rootDependencies, err := extractRootDependencies(environmentPackages, pkgName)
-	if err != nil {
-		return err
-	}
-	extractor.rootDependencies = rootDependencies
-
-	// Extract all project dependencies.
-	allDeps, childMap, err := extractDependencies(extractor.rootDependencies, environmentPackages)
-	if err != nil {
-		return err
-	}
-
-	// Update extracted dependencies.
-	extractor.allDependencies = allDeps
-	extractor.childrenMap = childMap
-
-	return nil
-}
-
-func (extractor *setupExtractor) PackageName() (string, error) {
-	var err error
-	extractor.once.Do(func() {
-		extractor.Pkg, err = getProjectName(extractor.pythonExecutablePath, extractor.setuppyFilePath)
-	})
-	return extractor.Pkg, err
-}
-
-func extractRootDependencies(envDeps map[string]pipDependencyPackage, pkgName string) ([]string, error) {
-	// Get installed package from environment-dependencies map.
-	pipDepPkg, ok := envDeps[strings.ToLower(pkgName)]
-	if !ok {
-		// Package not installed.
-		return nil, errorutils.CheckError(errors.New(fmt.Sprintf("Failed receiving root-dependencies for installed package: %s", pkgName)))
-	}
-
-	// Extract package's root-dependencies.
-	return pipDepPkg.getDependencies(), nil
-}
-
 // Get the project-name by running 'egg_info' command on setup.py and extracting it from 'PKG-INFO' file.
-func getProjectName(pythonExecutablePath, setuppyFilePath string) (string, error) {
+func ExtractPackageNameFromSetupPy(setuppyFilePath, pythonExecutablePath string) (string, error) {
 	// Execute egg_info command and return PKG-INFO content.
-	content, err := getEgginfoPkginfoContent(pythonExecutablePath, setuppyFilePath)
+	content, err := getEgginfoPkginfoContent(setuppyFilePath, pythonExecutablePath)
 	if err != nil {
 		return "", err
 	}
@@ -114,7 +43,7 @@ func getProjectNameFromFileContent(content []byte) (string, error) {
 
 // Run egg-info command on setup.py, the command generates metadata files.
 // Return the content of the 'PKG-INFO' file.
-func getEgginfoPkginfoContent(pythonExecutablePath, setuppyFilePath string) ([]byte, error) {
+func getEgginfoPkginfoContent(setuppyFilePath, pythonExecutablePath string) ([]byte, error) {
 	tempDirPath, err := fileutils.CreateTempDir()
 	if err != nil {
 		return nil, err
@@ -173,22 +102,10 @@ func extractPkginfoPathFromCommandOutput(egginfoOutput string) (string, error) {
 
 // Execute egg_info command for setup.py, return command's output.
 func executeEgginfoCommandWithOutput(pythonExecutablePath, setuppyFilePath string) (string, error) {
-	pythonEggInfoCmd := &pip.PipCmd{
+	pythonEggInfoCmd := &PipCmd{
 		Executable:  pythonExecutablePath,
 		Command:     setuppyFilePath,
 		CommandArgs: []string{"egg_info"},
 	}
 	return gofrogcmd.RunCmdOutput(pythonEggInfoCmd)
-}
-
-func (extractor *setupExtractor) AllDependencies() map[string]*buildinfo.Dependency {
-	return extractor.allDependencies
-}
-
-func (extractor *setupExtractor) DirectDependencies() []string {
-	return extractor.rootDependencies
-}
-
-func (extractor *setupExtractor) ChildrenMap() map[string][]string {
-	return extractor.childrenMap
 }
