+++ artifactory/utils/pip/setuppy_test.go
@@ -1,56 +1,8 @@
-package dependencies
+package pip
 
 import (
-	"fmt"
-	"reflect"
 	"testing"
 )
-
-var testDependencyMap = map[string]pipDependencyPackage{
-	"pkg1": {
-		packageType{"pkg1", "Pkg1", "1.0.0"},
-		[]dependency{
-			{"dep11", "Dep11", "1.0.1"},
-			{"dep12", "Dep12", "1.4.1"}}},
-	"pkg2": {
-		packageType{"pkg2", "PkG2", "2.2.2"},
-		[]dependency{
-			{"dep21", "DEP21", "2.0.1"},
-		}},
-	"pkg-.3": {
-		packageType{"pkg-.3", "PkG-.3", "3.10.1"},
-		[]dependency{
-			{"de--p31", "dE--P31", "3.0.15"},
-			{"dependency-of-3-2", "dEpenDenCy-oF-3-2", "3.3.1"},
-			{"d.e.p.3.3", "d.E.P.3.3", "30.0.1"},
-			{"d..ep34", "d..EP34", "343.334.443"},
-		}},
-	"pkg4": {
-		packageType{"pkg4", "PkG4", "4.2.4"},
-		[]dependency{},
-	}}
-
-func TestExtractRootDependencies(t *testing.T) {
-	tests := []struct {
-		packageName string
-		rootDeps    []string
-	}{
-		{"pkg1", []string{"dep11", "dep12"}},
-		{"pkg2", []string{"dep21"}},
-		{"pkg-.3", []string{"de--p31", "dependency-of-3-2", "d.e.p.3.3", "d..ep34"}},
-		{"pkg4", nil},
-	}
-
-	for _, test := range tests {
-		actualValue, err := extractRootDependencies(testDependencyMap, test.packageName)
-		if err != nil {
-			t.Error(err)
-		}
-		if !reflect.DeepEqual(actualValue, test.rootDeps) {
-			t.Error(fmt.Sprintf("Expected value: %v, got: %v.", test.rootDeps, actualValue))
-		}
-	}
-}
 
 func TestGetProjectNameFromFileContent(t *testing.T) {
 	tests := []struct {
