+++ build.bat
@@ -1,2 +1,1 @@
-python/addresources.bat
 CGO_ENABLED=0 go build -o jfrog.exe -ldflags '-w -extldflags "-static"' main.go
