+++ build.sh
@@ -7,6 +7,4 @@ if [ $# -eq 0 ]
 	exe_name="$1"
 fi
 
-./python/addresources.sh
-
 CGO_ENABLED=0 go build -o $exe_name -ldflags '-w -extldflags "-static"' main.go
