--- docs/artifactory/pipdepstree/help.go
@@ -1,8 +1,0 @@
-package pipdepstree
-
-const Description = "Show pip dependency tree."
-
-var Usage = []string{`jfrog rt pdt <pip command>`}
-
-const Arguments string = `	pip command
-		Arguments and options for the pip command.`
