+++ pip_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -13,8 +12,6 @@ import (
 	"testing"
 
 	gofrogcmd "github.com/jfrog/gofrog/io"
-	"github.com/jfrog/jfrog-cli-go/artifactory/commands/pip"
-	piputils "github.com/jfrog/jfrog-cli-go/artifactory/utils/pip"
 	"github.com/jfrog/jfrog-cli-go/inttestutils"
 	"github.com/jfrog/jfrog-cli-go/utils/cliutils"
 	"github.com/jfrog/jfrog-cli-go/utils/tests"
@@ -92,7 +89,7 @@ func testPipCmd(t *testing.T, outputFolder, projectPath, buildNumber, module str
 
 	err = artifactoryCli.Exec(args...)
 	if err != nil {
-		assert.Failf(t, "Failed executing pip-install command: %s", err.Error())
+		assert.Fail(t, "Failed executing pip-install command", err.Error())
 		cleanPipTest(t, outputFolder)
 		return
 	}
@@ -218,128 +215,4 @@ func (pfc *PipCmd) GetStdWriter() io.WriteCloser {
 
 func (pfc *PipCmd) GetErrWriter() io.WriteCloser {
 	return nil
-}
-
-func TestPipDepsTree(t *testing.T) {
-	initPipTest(t)
-
-	// Add virtual-environment path to 'PATH' for executing all pip and python commands inside the virtual-environment.
-	pathValue := setPathEnvForPipInstall(t)
-	if t.Failed() {
-		t.FailNow()
-	}
-	defer os.Setenv("PATH", pathValue)
-
-	// Check pip env is clean.
-	validateEmptyPipEnv(t)
-
-	// Populate cli config with 'default' server.
-	oldHomeDir, newHomeDir := prepareHomeDir(t)
-	defer os.Setenv(cliutils.HomeDir, oldHomeDir)
-	defer os.RemoveAll(newHomeDir)
-
-	// Create test cases.
-	allTests := []struct {
-		name                 string
-		project              string
-		outputFolder         string
-		moduleId             string
-		args                 []string
-		expectedDependencies int
-		cleanAfterExecution  bool
-	}{
-		{"setuppy", "setuppyproject", "setuppy", "jfrog-python-example", []string{".", "--no-cache-dir", "--force-reinstall"}, 3, true},
-		{"setuppy-verbose", "setuppyproject", "setuppy-verbose", "jfrog-python-example", []string{".", "--no-cache-dir", "--force-reinstall", "-v"}, 3, true},
-		{"setuppy-with-module", "setuppyproject", "setuppy-with-module", "setuppy-with-module", []string{".", "--no-cache-dir", "--force-reinstall"}, 3, true},
-		{"requirements", "requirementsproject", "requirements", tests.PipBuildName, []string{"-r", "requirements.txt", "--no-cache-dir", "--force-reinstall"}, 5, true},
-		{"requirements-verbose", "requirementsproject", "requirements-verbose", tests.PipBuildName, []string{"-r", "requirements.txt", "--no-cache-dir", "--force-reinstall", "-v"}, 5, false},
-		{"requirements-use-cache", "requirementsproject", "requirements-verbose", "requirements-verbose-use-cache", []string{"-r", "requirements.txt"}, 5, true},
-	}
-
-	// Run test cases.
-	for _, test := range allTests {
-		t.Run(test.name, func(t *testing.T) {
-			testPipDepsTreeCmd(t, createPipProject(t, test.outputFolder, test.project), test.expectedDependencies, test.args)
-			if test.cleanAfterExecution {
-				// cleanup
-				cleanPipTest(t, test.name)
-			}
-		})
-	}
-	cleanPipTest(t, "cleanup")
-	tests.CleanFileSystem()
-}
-
-func testPipDepsTreeCmd(t *testing.T, projectPath string, expectedElements int, args []string) {
-	wd, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = os.Chdir(projectPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Chdir(wd)
-
-	// Get pip configuration.
-	pipConfig, err := piputils.GetPipConfiguration()
-	if err != nil {
-		t.Fatalf("Error occurred while attempting to read pip-configuration file: %s\n"+
-			"Please run 'jfrog rt pip-config' command prior to running 'jfrog rt pip-deps-tree'.", err.Error())
-	}
-	// Set arg values.
-	rtDetails, err := pipConfig.RtDetails()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create command.
-	pipDepsTreeCmd := pip.NewPipDepTreeCommand()
-	pipDepsTreeCmd.SetRtDetails(rtDetails).SetRepo(pipConfig.TargetRepo()).SetArgs(args)
-
-	err = pipDepsTreeCmd.Run()
-	if err != nil {
-		t.Fatalf("Failed while executing pip-deps-tree command: %s", err)
-	}
-
-	// Check result elements.
-	treeJsonData, err := pipDepsTreeCmd.DepsTreeRoot.MarshalJSON()
-	if err != nil {
-		t.Fatalf("Failed parsing tree json: %s", err)
-	}
-
-	// Count dependencies.
-	var depsTreeTest []DependenciesTreeTest
-	err = json.Unmarshal(treeJsonData, &depsTreeTest)
-	assert.NoError(t, err)
-	depsCount := countDependencies(depsTreeTest)
-	assert.Equal(t, expectedElements, depsCount, "Incorrect number of dependencies found")
-}
-
-type DependenciesTreeTest struct {
-	Id                 string                 `json:"id,omitempty"`
-	DirectDependencies []DependenciesTreeTest `json:"dependencies,omitempty"`
-}
-
-func countDependencies(allDeps []DependenciesTreeTest) int {
-	depsMap := make(map[string]int)
-	// Iterate over dependencies, resolve and discover more dependencies.
-	index := -1
-	var currentDep string
-	for {
-		index++
-		// Check if should stop.
-		if len(allDeps) < index+1 {
-			break
-		}
-		currentDep = allDeps[index].Id
-		// Check if current dependency already resolved.
-		if _, ok := depsMap[currentDep]; ok {
-			// Already resolved.
-			continue
-		}
-		// Add currentDep dependencies for cound.
-		allDeps = append(allDeps, allDeps[index].DirectDependencies...)
-	}
-	return index
 }
