--- python/addresources.sh
@@ -1,1 +1,0 @@
-go run ./python/addresources.go
