--- test.sh
@@ -1,2 +1,0 @@
-./python/addresources.sh
-go test $@
